Cleaned up and refactored code and split classes in different files
