+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,14 +1,16 @@
 ---
 name: Feature request
 about: Suggest an idea for this project
-
 ---
 
-**Describe the feature you would like**
-A clear and concise description of what you want to happen.
+## Description
 
-**Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+<!--- A clear and concise description of what you want to happen. -->
 
-**Additional context**
-Add any other context or screenshots about the feature request here.
+## Problem
+
+<!--- A clear and concise description of what the problem is. e.g.: I'm always frustrated when [...] -->
+
+## Additional context
+
+<!--- Add any other context or screenshots about the feature request here. -->
