+++ cryptic/errors.py
@@ -1,0 +1,18 @@
+class IllegalReturnTypeError(Exception):
+    pass
+
+
+class UnknownDBMSTypeError(ValueError):
+    pass
+
+
+class UnknownModeError(ValueError):
+    pass
+
+
+class FrameTooLongError(Exception):
+    pass
+
+
+class FrameCorruptedError(Exception):
+    pass
